+++ lib/osm/event.rb
@@ -14,7 +14,7 @@ module Osm
     # @!attribute [rw] finish
     #   @return [DateTime] when the event ends
     # @!attribute [rw] cost
-    #   @return [String] the cost of the event
+    #   @return [String] the cost of the event (formatted to \d+\.\d{2}) or "TBC"
     # @!attribute [rw] location
     #   @return [String] where the event is
     # @!attribute [rw] notes
@@ -45,7 +45,7 @@ module Osm
     attribute :name, :type => String
     attribute :start, :type => DateTime
     attribute :finish, :type => DateTime
-    attribute :cost, :type => String, :default => ''
+    attribute :cost, :type => String, :default => 'TBC'
     attribute :location, :type => String, :default => ''
     attribute :notes, :type => String, :default => ''
     attribute :archived, :type => Boolean, :default => false
@@ -72,6 +72,7 @@ module Osm
     validates_inclusion_of :reminders, :in => [true, false]
     validates_inclusion_of :attendance_limit_includes_leaders, :in => [true, false]
     validates_inclusion_of :allow_booking, :in => [true, false]
+    validates_format_of :cost, :with => /\A(?:\d+\.\d{2}|TBC)\Z/
 
 
     # @!method initialize
@@ -153,7 +154,7 @@ module Osm
         'location' => event.location,
         'startdate' => event.start? ? event.start.strftime(Osm::OSM_DATE_FORMAT) : '',
         'enddate' => event.finish? ? event.finish.strftime(Osm::OSM_DATE_FORMAT) : '',
-        'cost' => event.cost,
+        'cost' => event.cost_tbc? ? '-1' : event.cost,
         'notes' => event.notes,
         'starttime' => event.start? ? event.start.strftime(Osm::OSM_TIME_FORMAT) : '',
         'endtime' => event.finish? ? event.finish.strftime(Osm::OSM_TIME_FORMAT) : '',
@@ -191,7 +192,7 @@ module Osm
         'location' => location,
         'startdate' => start? ? start.strftime(Osm::OSM_DATE_FORMAT) : '',
         'enddate' => finish? ? finish.strftime(Osm::OSM_DATE_FORMAT) : '',
-        'cost' => cost,
+        'cost' => cost_tbc? ? '-1' : cost,
         'notes' => notes,
         'starttime' => start? ? start.strftime(Osm::OSM_TIME_FORMAT) : '',
         'endtime' => finish? ? finish.strftime(Osm::OSM_TIME_FORMAT) : '',
@@ -339,6 +340,18 @@ module Osm
       return attendance_limit - attendees(api)
     end
 
+    # Whether the cost is to be confirmed
+    # @return [Boolean] whether the cost is TBC
+    def cost_tbc?
+      cost.eql?('TBC')
+    end
+
+    # Whether the cost is zero
+    # @return [Boolean] whether the cost is zero
+    def cost_free?
+      cost.eql?('0.00')
+    end
+
     # Compare Event based on start, name then id
     def <=>(another)
       return 0 if self.id == another.try(:id)
@@ -365,7 +378,7 @@ module Osm
         :name => event_data['name'],
         :start => Osm::make_datetime(event_data['startdate'], event_data['starttime']),
         :finish => Osm::make_datetime(event_data['enddate'], event_data['endtime']),
-        :cost => event_data['cost'],
+        :cost => event_data['cost'].eql?('-1') ? 'TBC' : event_data['cost'],
         :location => event_data['location'],
         :notes => event_data['notes'],
         :archived => event_data['archived'].eql?('1'),